backend/requirements.txt
# W&M Course Catalog Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Async HTTP Client
aiohttp>=3.9.0

# Firebase
firebase-admin>=6.2.0

# Environment Variables
python-dotenv>=1.0.0

# Redis Cache (optional)
redis>=5.0.0

# Task Scheduler
apscheduler>=3.10.0

# HTTP Requests (for curriculum scraper)
requests>=2.31.0

# PDF Parsing (for curriculum scraper)
pdfplumber>=0.10.0

# Gaze Tracking / Heatmap Generation (debug mode)
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
Pillow>=10.0.0

# --- Development/Testing Dependencies ---
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
freezegun>=1.2.0
httpx>=0.25.0
